refactor: split up cmd package files
