--- cmd/kconf.go
@@ -1,236 +1,0 @@
-package cmd
-
-import (
-	"errors"
-	"fmt"
-	"os"
-
-	"github.com/rs/zerolog"
-
-	"github.com/rs/zerolog/log"
-	"github.com/spf13/cobra"
-
-	"github.com/particledecay/kconf/build"
-	"github.com/particledecay/kconf/pkg/kubeconfig"
-)
-
-var (
-	verbose     bool
-	contextName string
-)
-
-var rootCmd = &cobra.Command{
-	Use:   "kconf",
-	Short: "kconf manages your kubeconfigs",
-	Long: `kconf allows you to add and delete kubeconfigs by merging kubeconfig
-			files together and breaking them apart appropriately.`,
-	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-		if verbose {
-			zerolog.SetGlobalLevel(zerolog.DebugLevel)
-		}
-	},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if len(args) < 1 {
-			return errors.New("An action is required")
-		}
-		return nil
-	},
-}
-
-var addCmd = &cobra.Command{
-	Use:     "add",
-	Short:   "Add in a new kubeconfig file and optional context name",
-	Long:    `Add a new kubeconfig file to the existing merged config file and optional context name`,
-	Aliases: []string{"a"},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if len(args) < 1 {
-			return errors.New("You must supply the path to a kubeconfig file")
-		}
-		return nil
-	},
-	Run: func(cmd *cobra.Command, args []string) {
-		filepath := args[0]
-		config, err := kubeconfig.GetConfig()
-		if err != nil {
-			log.Fatal().Msgf("Error while reading main config: %v", err)
-		}
-
-		newConfig, err := kubeconfig.Read(filepath)
-		if err != nil {
-			log.Fatal().Msgf("Error while reading %s: %v", filepath, err)
-		}
-		if config == nil {
-			log.Fatal().Msgf("Could not find kubeconfig at %s", filepath)
-		}
-
-		err = config.Merge(newConfig, contextName)
-		if err != nil {
-			log.Fatal().Msgf("%v", err)
-		}
-		err = config.Save()
-		if err != nil {
-			log.Fatal().Msgf("%v", err)
-		}
-	},
-}
-
-var removeCmd = &cobra.Command{
-	Use:     "rm",
-	Short:   "Remove a kubeconfig from main file",
-	Long:    `Remove a named context and associated resources from main kubeconfig file`,
-	Aliases: []string{"r"},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if len(args) < 1 {
-			return errors.New("You must provide the name of a kubeconfig context")
-		}
-		return nil
-	},
-	Run: func(cmd *cobra.Command, args []string) {
-		contextName := args[0]
-		config, err := kubeconfig.GetConfig()
-		if err != nil {
-			log.Fatal().Msgf("Error while reading main config: %v", err)
-		}
-		err = config.Remove(contextName)
-		if err != nil {
-			log.Fatal().Msgf("%v", err)
-		}
-		err = config.Save()
-		if err != nil {
-			log.Fatal().Msgf("%v", err)
-		}
-	},
-}
-
-var listCmd = &cobra.Command{
-	Use:     "list",
-	Short:   "List all saved contexts",
-	Long:    `Print a list of all contexts previously saved in kubeconfig`,
-	Aliases: []string{"ls"},
-	Args:    cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		config, err := kubeconfig.GetConfig()
-		if err != nil {
-			log.Fatal().Msgf("Could not read main config")
-		}
-		contexts, currentContext := config.List()
-		for _, ctx := range contexts {
-			if currentContext == ctx {
-				fmt.Println("*", ctx)
-			} else {
-				fmt.Println(" ", ctx)
-			}
-		}
-	},
-}
-
-var viewCmd = &cobra.Command{
-	Use:     "view",
-	Short:   "View a specific context's config",
-	Long:    `Display all of the config resources associated with a specific context`,
-	Aliases: []string{"v"},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if len(args) < 1 {
-			return errors.New("You must provide the name of a kubeconfig context")
-		}
-		return nil
-	},
-	Run: func(cmd *cobra.Command, args []string) {
-		contextName := args[0]
-		config, err := kubeconfig.GetConfig()
-		if err != nil {
-			log.Fatal().Msg("Could not read main config")
-		}
-
-		// convert config into bytes
-		content, err := config.GetContent(contextName)
-		if err != nil {
-			log.Fatal().Msgf("Error while converting context '%s': %v", contextName, err)
-		}
-
-		// print config content
-		os.Stdout.Write(content)
-	},
-}
-
-var useCmd = &cobra.Command{
-	Use:     "use",
-	Short:   "Set the current context",
-	Long:    `Set the current context in the main kubeconfig`,
-	Aliases: []string{"u"},
-	Args: func(cmd *cobra.Command, args []string) error {
-		if len(args) < 1 {
-			return errors.New("You must provide the name of a kubeconfig context")
-		}
-		return nil
-	},
-	Run: func(cmd *cobra.Command, args []string) {
-		contextName := args[0]
-
-		config, err := kubeconfig.GetConfig()
-		if err != nil {
-			log.Fatal().Msgf("Could not read main config")
-		}
-
-		contexts, currentContext := config.List()
-		if currentContext == contextName {
-			fmt.Printf("Current context is already '%s'\n", currentContext)
-			return
-		}
-
-		hasContext := false
-		for _, ctx := range contexts {
-			if contextName == ctx {
-				hasContext = true
-				break
-			}
-		}
-		if hasContext {
-			config.WriteCurrentContext(contextName)
-			config.Save()
-			if err == nil {
-				fmt.Printf("Using context '%s'\n", contextName)
-			} else {
-				log.Fatal().Msgf("Could not save current context to main config")
-			}
-		} else {
-			fmt.Printf("Context '%s' not found\n", contextName)
-		}
-	},
-}
-
-var versionCmd = &cobra.Command{
-	Use:     "version",
-	Short:   "Print version",
-	Long:    `Print version information`,
-	Aliases: []string{"ver"},
-	Args:    cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		if verbose == true {
-			if err := build.PrintLongVersion(); err != nil {
-				log.Error().Msgf("%v", err)
-			}
-		} else {
-			build.PrintVersion()
-		}
-	},
-}
-
-func init() {
-	// flags
-	rootCmd.PersistentFlags().BoolVarP(&verbose, "verbose", "v", false, "display debug messages")
-	addCmd.Flags().StringVarP(&contextName, "context-name", "n", "", "override context name")
-}
-
-// Execute combines all of the available command functions
-func Execute() {
-	rootCmd.AddCommand(addCmd)
-	rootCmd.AddCommand(removeCmd)
-	rootCmd.AddCommand(listCmd)
-	rootCmd.AddCommand(viewCmd)
-	rootCmd.AddCommand(useCmd)
-	rootCmd.AddCommand(versionCmd)
-	if err := rootCmd.Execute(); err != nil {
-		log.Fatal().Msgf("Error during execution: %v", err)
-	}
-}
